requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Limiter==3.4.1
python-dotenv==1.0.1
